Match sub_8023B5C

+++ include/game/stage/player.h
@@ -15,7 +15,7 @@ void Player_ClearMovestate_IsInScriptedSequence(void);
 
 void InitializePlayer(Player *p);
 void sub_80218E4(Player *);
-void sub_8023B5C(Player *, s8);
+void sub_8023B5C(Player *, s32);
 void sub_8023260(Player *);
 void sub_80232D0(Player *);
 void sub_8023610(Player *);
